refactor(InputCurrency): Change default currency to IDR
This commit updates the hardcoded default currency settings within the `InputCurrency` component.

The configuration for the `useCurrencyInput` hook is changed from Euro (EUR) with a German locale (`de-DE`) to Indonesian Rupiah (IDR) with an Indonesian locale (`id-ID`).

Additionally, the `currencyDisplay` is set from `hidden` to `symbol` to make the currency explicit in the input field.